GODRIVER-2911: Move things as necessary for authentication registration
