+++ x/mongo/driver/auth/cred.go
@@ -3,14 +3,10 @@
 // Licensed under the Apache License, Version 2.0 (the "License"); you may
 // not use this file except in compliance with the License. You may obtain
 // a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-
 package auth
 
-// Cred is a user's credential.
-type Cred struct {
-	Source      string
-	Username    string
-	Password    string
-	PasswordSet bool
-	Props       map[string]string
-}
+import (
+	"go.mongodb.org/mongo-driver/x/mongo/driver"
+)
+
+type Cred = driver.AuthCred
